GH-46: Extend pattern for XdtSource + XdtTransformationSettings
This shifts the way in which we pass the "transform"
document/specification into the transformation methods & aliases, as
well as the way in which the various logging implementations are
being exposed/consumed.

By introducing the XdtTransform type and forcing the
objects to expose a stream, we can support a more
flexible way of expressing the transformation -
two such examples are an in-memory/streamed document
or an XML fragment, which can be useful when you only need
to express a simple <appSettings> transformation of a line or two,
or if the transform itself is built at runtime.

To address the various named overloads to facilitate the
logging pattersn, this change shifts that concept and converges
the IXmlTransformationLogger managment toward a
XdtTransformationSettings type.

The changes made leave all existing methods and aliases in place for
compatibility and conforms to existing patterns -
a null logger will throw exceptions (from within the DotNet.Xdt
library itself), and the 'default' logger is still considered
to be XdtTransformationLog.  The one change to consider was the
shift of XdtTransformation from 'static' to 'public sealed', but
this is not a breaking change nor a concern.  Accepting the
IFileSystem as a ctor parameter is both necessary and a stronger design.

The one thing to note is the fact that the XdtFileSource is
internal; rather than force a caller to make an awkward transition
from FilePath -> IFile, the alias(es) exposed for a transform FilePath
specification will provide this support.  Rather than expose a
FilePathXdtSource sort of object (requiring type checking and
adaptation to support the GetStream method), this should be handled
by the implementation, making it friendlier for the developer/end-user.

This also addressed a couple of unrelated issues:
 - The *.resx 'Resources' objects - at some point the custom tool
   specification had been lost and the *.Designer.cs file no longer
   updated with changes to the *.resx; this was corrected.
 - Various typos have been corrected